refactor: Make p2p example start the session in a system
This is more similar to what would be done in an actual game. i.e. you
don't necessarily know what peers to connect to prior to the game
launch, so you would probably start the session when entering a new
state. In any case, this makes the example easier to adapt to real use
cases, while still keeping it relatively simple.