feat: Add a way to stop sessions
i.e.

commands.stop_session()

With this change it's possible to stop and start a new session without
recreating the bevy app.

The implementation is messy (with the marker resource), but it was the
easiest thing to do that I could think of at the moment. At least it's
private.