Updated Rollback Component to use Bevy Entity
Replaced the existing `Rollback` component with a new component `RollbackFlag` which contains an `Entity` as its discriminant instead of a `u32`.

Replaced `RollbackIdProvider` with an `EntityCommand` `Rollback` which injects the appropriate `RollbackFlag` into an entity.

Updated tests and examples to reflect changes.

Moved `Rollback` related functionality into a separate module for maintainability.